requirements.txt
aiohttp
async-timeout
attrs
certifi
chardet
discord.py
dnspython
idna
multidict
pymongo
requests
urllib3
websockets
yarl
